2023/8/8

+++ src/Contests/Huyue/2023-7/B.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 struct node {
     unsigned long long value;
-    int next;
+    long long next;
 };
 
 node nodes[10005];
@@ -14,14 +14,14 @@ int main() {
     // freopen("song.out", "w", stdout);
     unsigned long long n, t;
     cin >> n >> t;
-    for (int i = 1; i <= n; i++) {
-        int a;
+    for (long long i = 1; i <= n; i++) {
+        long long a;
         cin >> a;
         nodes[i].value = a;
         nodes[i].next = i + 1;
     }
     nodes[n].next = 1;
-    int tmp = 1;
+    long long tmp = 1;
     while (1) {
         if (t < nodes[tmp].value) {
             cout << tmp << ' ' << t << endl;
